refactor(emit): extract backup subsystem → backup.py + atomicio.py (CH-07)
Pure internal reorganization (move + re-export; zero behavior change) to stop emit.py's
accretion and isolate the data-safety backup subsystem that grew it 1389 → 1584.

- agentteams/atomicio.py — leaf module: the atomic write/copy primitives
  (_target_mode, _atomic_write_text, _atomic_copy) + _resolve_path. backup imports
  from here (not emit), keeping the graph acyclic: atomicio ← backup ← emit.
- agentteams/backup.py — backup / restore / verify / prune / off-machine mirror
  (BackupResult, PruneResult, backup_output_dir, list_backups, verify_backup,
  prune_backups, restore_backup, _mirror_backup, manifest helpers, constants).
- emit.py re-exports every moved public + private symbol (39 external call sites use
  emit.backup_output_dir / emit._atomic_write_text / etc.), so callers resolve them
  unchanged. emit.py: 1584 → 1080.

Two test_emit.py atomicity tests repointed to monkeypatch atomicio.os.replace (os.replace
now executes in atomicio). LENGTH_ALLOWLIST comment refreshed (1584 → 1080); atomicio.py +
backup.py added to _REFACTOR_MODULES so the CH-22 annotation ratchet covers them.

Plan adversarial+conflict audited — the adversarial auditor performed the full extraction
in a scratch copy and ran the suite (zero regressions). No CHANGELOG entry (pure reorg,
per the cli/ decomposition precedent). Full suite 1385 passed; man/RSR1/hygiene green.
