+++ tests/test_code_hygiene.py
@@ -38,10 +38,12 @@ MAX_MODULE_LINES = 1000
 LENGTH_ALLOWLIST: frozenset[str] = frozenset({
     # build_team.py left at Step D (now a 833-line shim); cli/app.py left at Step D2
     # (1174 -> 263 after the generate pipeline moved to cli/generate.py, 939 lines).
-    "agentteams/analyze.py",    # 1504 — accepted tracked debt (CH-07 allowlist; not split)
-    "agentteams/emit.py",       # 1584 — accepted tracked debt (CH-07 allowlist; not split).
-    # NOTE: emit.py is accreting (1389 -> 1584 after backup retention/mirror + verify).
-    # Standing revisit-trigger: split it the next time a feature must touch it heavily.
+    "agentteams/analyze.py",    # 1507 — accepted tracked debt (CH-07 allowlist; not split)
+    "agentteams/emit.py",       # 1080 — accepted tracked debt (CH-07 allowlist; not split).
+    # The backup subsystem (was driving emit.py's accretion to 1584) was extracted to
+    # agentteams/backup.py + agentteams/atomicio.py (CH-07), bringing emit.py 1584 -> 1080.
+    # Getting under the 1000 ceiling would need a second carve (fence/merge -> fences.py);
+    # deferred. analyze.py still pending its own carve (_plan_output_files -> output_plan.py).
 })
 BROAD_EXCEPT_BASELINE = 11      # except Exception/BaseException/bare. Narrowed over the sweep
                                 # (Steps E + remaining-items I6: commands, render_pipeline, ingest,
@@ -198,6 +200,9 @@ _REFACTOR_MODULES = (
     "agentteams/cli/security_gate.py",
     "agentteams/frameworks/registry.py",
     "agentteams/errors.py",
+    # CH-07 module extractions (keep the annotation ratchet honest on new modules)
+    "agentteams/atomicio.py",
+    "agentteams/backup.py",
 )
